feat: 패키지 경로 변경 및 entity - domain 변환 로직 추가
